wip on horrific code
